refactor: validations in delete methods

+++ src/main/java/project/hotel_reservations/service/GuestServiceImpl.java
@@ -95,6 +95,10 @@ public class GuestServiceImpl implements GuestService {
     @Override
     @Transactional
     public void delete(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("User not found");
+        }
+
         repository.deleteById(id);
     }
 }
